requirements.txt
langgraph>=0.2.60
langgraph-checkpoint-sqlite>=2.0.0
langchain-core>=0.3.0
langchain-google-genai>=2.0.0
pydantic>=2.7.0
streamlit>=1.38.0
python-dotenv>=1.0.1
pytest>=8.0.0
